refactor: extract cli utilities
