+++ internal/cliutil/doc.go
@@ -1,0 +1,3 @@
+// Package cliutil contains shared helpers for CLI commands such as stack
+// loading and log formatting.
+package cliutil
